requirements.txt
altair==5.3.0
annotated-types==0.6.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==4.4.2
docstring_parser==0.16
executing==2.0.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.21.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.2
grpcio-status==1.62.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain-core==0.1.46
langchain-google-vertexai==1.0.1
langsmith==0.1.51
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==1.0.3
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.1
packaging==23.2
pandas==2.2.2
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0b1
pydub==0.25.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
referencing==0.35.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
shapely==2.0.4
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.33.0
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
types-protobuf==4.25.0.20240417
types-requests==2.31.0.20240406
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vertexai==1.48.0
wcwidth==0.2.13
